requirements.txt
aiobotocore==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aioitertools==0.13.0
aiosignal==1.4.0
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
appdirs==1.4.4
asyncssh==2.22.0
atpublic==7.0.0
attrs==26.1.0
backoff==2.2.1
beautifulsoup4==4.14.3
billiard==4.2.4
blinker==1.9.0
boto3==1.42.84
botocore==1.42.84
cachetools==5.5.2
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
configobj==5.0.9
contourpy==1.3.3
cryptography==43.0.3
cycler==0.12.1
dacite==1.6.0
dagshub==0.7.0
dagshub-annotation-converter==0.2.0
databricks-sdk==0.102.0
dataclasses-json==0.6.7
dictdiffer==0.9.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docker==7.1.0
dpath==2.2.0
dulwich==0.24.1
dvc==3.67.1
dvc-data==3.18.3
dvc-gdrive==3.0.1
dvc-http==2.32.0
dvc-objects==5.2.0
dvc-render==1.0.2
dvc-s3==3.3.0
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
entsoe-py==0.8.0
fastapi==0.136.0
filelock==3.29.0
flake8==7.3.0
Flask==3.1.3
flask-cors==6.0.2
flatten-dict==0.4.2
flufl.lock==9.0.0
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.3.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
googleapis-common-protos==1.74.0
gql==4.0.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
gto==1.9.0
gunicorn==23.0.0
h11==0.16.0
holidays==0.94
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huey==2.6.0
hydra-core==1.3.2
identify==2.6.19
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
kiwisolver==1.5.0
kombu==5.6.2
lxml==6.1.0
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mccabe==0.7.0
mdurl==0.1.2
mlflow==3.11.1
mlflow-skinny==3.11.1
mlflow-tracing==3.11.1
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
nodeenv==1.10.0
numpy==2.4.4
oauth2client==4.1.3
omegaconf==2.3.0
opentelemetry-api==1.40.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
optuna==4.8.0
orjson==3.11.8
packaging==24.2
pandas==2.3.3
pathspec==1.0.4
pathvalidate==3.3.1
pillow==12.2.0
platformdirs==4.9.6
pluggy==1.6.0
pre_commit==4.5.1
prettytable==3.17.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
psutil==7.2.2
pyarrow==18.1.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.13.2
pydantic-settings==2.13.1
pydantic_core==2.46.2
pydot==4.0.1
PyDrive2==1.21.3
pyflakes==3.4.0
pygit2==1.19.2
Pygments==2.20.0
pygtrie==2.5.0
pyOpenSSL==24.2.1
pyparsing==3.3.2
pytest==9.0.3
python-dateutil==2.9.0.post0
python-discovery==1.2.2
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
rsa==4.9.1
ruamel.yaml==0.19.1
ruff==0.15.11
s3fs==2026.3.0
s3transfer==0.16.0
scikit-learn==1.8.0
scipy==1.17.1
scmrepo==3.5.8
seaborn==0.13.2
semver==3.0.4
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.17.0
skops==0.13.0
smmap==5.0.3
soupsieve==2.8.3
SQLAlchemy==2.0.49
sqlparse==0.5.5
sqltrie==0.11.2
starlette==1.0.0
tabulate==0.10.0
tenacity==9.1.4
threadpoolctl==3.6.0
tomlkit==0.14.0
tqdm==4.67.3
treelib==1.8.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==1.26.20
uvicorn==0.44.0
vine==5.1.0
virtualenv==21.2.4
voluptuous==0.16.0
wcwidth==0.6.0
Werkzeug==3.1.8
wrapt==2.1.2
xgboost==3.2.0
yarl==1.23.0
zc.lockfile==4.0
zipp==3.23.0
